+++ lib/flame/runner.ex
@@ -49,7 +49,7 @@ defmodule FLAME.Runner do
             backend_init: nil,
             node_name: nil,
             single_use: false,
-            timeout: 20_000,
+            timeout: 30_000,
             status: nil,
             log: :info,
             boot_timeout: 10_000,
@@ -118,11 +118,10 @@ defmodule FLAME.Runner do
   def call(runner_pid, caller_pid, func, opts \\ [])
       when is_pid(runner_pid) and is_pid(caller_pid) and is_function(func) and is_list(opts) do
     link? = Keyword.get(opts, :link, true)
-    timeout = opts[:timeout] || nil
     track_resources? = Keyword.get(opts, :track_resources, false)
     {ref, %Runner{} = runner, backend_state} = checkout(runner_pid)
     %Runner{terminator: terminator} = runner
-    call_timeout = timeout || runner.timeout
+    call_timeout = opts[:timeout] || runner.timeout
 
     result =
       remote_call(runner, backend_state, call_timeout, track_resources?, fn ->
@@ -132,12 +131,12 @@ defmodule FLAME.Runner do
       end)
 
     case result do
-      {:ok, value} ->
-        :ok = checkin(runner_pid, ref)
-        value
+      {:ok, {value, trackable_pids}} ->
+        :ok = checkin(runner_pid, ref, trackable_pids)
+        {value, trackable_pids}
 
       {kind, reason} ->
-        :ok = checkin(runner_pid, ref)
+        :ok = checkin(runner_pid, ref, [])
 
         case kind do
           :exit -> exit(reason)
@@ -151,8 +150,8 @@ defmodule FLAME.Runner do
     GenServer.call(runner_pid, :checkout)
   end
 
-  defp checkin(runner_pid, ref) do
-    GenServer.call(runner_pid, {:checkin, ref})
+  defp checkin(runner_pid, ref, trackable_pids) do
+    GenServer.call(runner_pid, {:checkin, ref, trackable_pids})
   end
 
   @impl true
@@ -253,8 +252,6 @@ defmodule FLAME.Runner do
   end
 
   def handle_call(:checkout, {from_pid, _tag}, state) do
-    ref = Process.monitor(from_pid)
-
     state =
       case maybe_diff_code_paths(state) do
         {new_state, nil} ->
@@ -270,12 +267,20 @@ defmodule FLAME.Runner do
           new_state
       end
 
-    {:reply, {ref, state.runner, state.backend_state}, put_checkout(state, from_pid, ref)}
+    {new_state, ref} = put_checkout(state, from_pid)
+    {:reply, {ref, new_state.runner, new_state.backend_state}, new_state}
   end
 
-  def handle_call({:checkin, ref}, _from, state) do
+  def handle_call({:checkin, ref, trackable_pids}, _from, state) do
     Process.demonitor(ref, [:flush])
-    {:reply, :ok, drop_checkout(state, ref)}
+
+    new_state =
+      Enum.reduce(trackable_pids, state, fn pid, acc ->
+        {acc, _ref} = put_checkout(acc, pid)
+        acc
+      end)
+
+    {:reply, :ok, drop_checkout(new_state, ref)}
   end
 
   def handle_call({:remote_boot, base_sync_stream, _timeout}, _from, state) do
@@ -301,16 +306,17 @@ defmodule FLAME.Runner do
                 terminator: term
               } = new_runner
 
-              :ok =
+              {:ok, _} =
                 remote_call!(runner, new_backend_state, runner.boot_timeout, false, fn ->
                   # ensure app is fully started if parent connects before up
                   if otp_app, do: {:ok, _} = Application.ensure_all_started(otp_app)
 
+                  if base_sync_stream, do: CodeSync.extract_packaged_stream(base_sync_stream)
+                  if beams_stream, do: CodeSync.extract_packaged_stream(beams_stream)
+
                   :ok =
                     Terminator.schedule_idle_shutdown(term, idle_after, idle_check, single_use)
 
-                  if base_sync_stream, do: CodeSync.extract_packaged_stream(base_sync_stream)
-                  if beams_stream, do: CodeSync.extract_packaged_stream(beams_stream)
                   :ok
                 end)
 
@@ -342,12 +348,16 @@ defmodule FLAME.Runner do
       ])
 
     Keyword.validate!(opts[:code_sync] || [], [
+      :get_path,
+      :copy_apps,
       :copy_paths,
       :sync_beams,
       :start_apps,
       :tmp_dir,
       :extract_dir,
-      :verbose
+      :verbose,
+      :compress,
+      :chunk_size
     ])
 
     {idle_shutdown_after_ms, idle_check} =
@@ -404,8 +414,9 @@ defmodule FLAME.Runner do
     result
   end
 
-  defp put_checkout(state, from_pid, ref) when is_pid(from_pid) do
-    %{state | checkouts: Map.put(state.checkouts, ref, from_pid)}
+  defp put_checkout(state, from_pid) when is_pid(from_pid) do
+    ref = Process.monitor(from_pid)
+    {%{state | checkouts: Map.put(state.checkouts, ref, from_pid)}, ref}
   end
 
   defp drop_checkout(state, ref) when is_reference(ref) do
@@ -460,9 +471,9 @@ defmodule FLAME.Runner do
         if track_resources? do
           {result, pids} = FLAME.track_resources(result, [], node(remote_pid))
           send(remote_pid, {parent_ref, pids})
-          {:ok, result}
+          {:ok, {result, pids}}
         else
-          {:ok, result}
+          {:ok, {result, []}}
         end
 
       {:DOWN, ^remote_monitor_ref, :process, ^remote_pid, reason} ->
@@ -505,7 +516,11 @@ defmodule FLAME.Runner do
 
   defp maybe_stream_code_paths(%{runner: %Runner{} = runner} = state) do
     if code_sync_opts = runner.code_sync_opts do
-      code_sync = CodeSync.new(Keyword.put(code_sync_opts, :copy_paths, []))
+      code_sync =
+        code_sync_opts
+        |> CodeSync.new()
+        |> CodeSync.compute_sync_beams()
+
       %CodeSync.PackagedStream{} = parent_stream = CodeSync.package_to_stream(code_sync)
       new_runner = %Runner{runner | code_sync: code_sync}
       {%{state | runner: new_runner}, parent_stream}